+++ changelog.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 - Updates for unknown alarms for 4.5L controller
 - Updates to About page to better handle slower upgrades
 - Minor update to gensnmp to allow the disabling of IPv6
+- Added new add on program gencallmebot (thanks @buschauer)
+- Added preliminary support for PowerZone 410 controller
 
 ## V1.18.16 -2022-12-06
 - Change to allow CPU temp gauge for any linux system